fix: bug fix, refctor , prod flyway

+++ backend/src/main/java/com/nabinrai/futsal_team_manager/player/service/PlayerServiceImpl.java
@@ -129,6 +129,9 @@ public class PlayerServiceImpl implements PlayerService {
                                 "Player not found with id: " + id
                         )
                 );
+        if (Role.ADMIN.equals(player.getRole())) {
+            throw new IllegalArgumentException("Cannot delete an Admin player");
+        }
         matchParticipationRepository.deleteByPlayerId(id);
         playerRepository.delete(player);
